Related work items: #1053, #2179, #2356, #2391, #2408, #2409, #2435, #2457, #2544, #2551, #2566, #2589, #2601, #2609, #2610, #2611, #2613, #2617, #2621, #2623, #2626, #2627, #2628, #2650

+++ Common/DataModels/Item.cs
@@ -26,7 +26,7 @@ namespace Ferretto.Common.DataModels
 
         public int? FifoTimeStore { get; set; }
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
         public int Id { get; set; }
 
@@ -52,7 +52,7 @@ namespace Ferretto.Common.DataModels
 
         public DateTime? LastStoreDate { get; set; }
 
-        public int? Length { get; set; }
+        public double? Length { get; set; }
 
         public ItemManagementType ManagementType { get; set; }
 
@@ -74,7 +74,7 @@ namespace Ferretto.Common.DataModels
 
         public int? StoreTolerance { get; set; }
 
-        public int? Width { get; set; }
+        public double? Width { get; set; }
 
         #endregion
     }
