+++ WMS/Data/Core/Interfaces/ICompartmentProvider.cs
@@ -17,7 +17,7 @@ namespace Ferretto.WMS.Data.Core.Interfaces
         #region Methods
 
         Task<IOperationResult<IEnumerable<CompartmentDetails>>> CreateRangeAsync(
-            IEnumerable<CompartmentDetails> compartments);
+                    IEnumerable<CompartmentDetails> compartments);
 
         Task<IEnumerable<AllowedItemInCompartment>> GetAllowedItemsAsync(int id);
 
@@ -25,7 +25,7 @@ namespace Ferretto.WMS.Data.Core.Interfaces
 
         Task<IEnumerable<CompartmentDetails>> GetByLoadingUnitIdAsync(int id);
 
-        Task<int?> GetMaxCapacityAsync(int width, int height, int itemId);
+        Task<double?> GetMaxCapacityAsync(double width, double height, int itemId);
 
         #endregion
     }
